Divided login page to components
