+++ src/pages/Login.vue
@@ -8,113 +8,20 @@
             <div class="col-auto self-center q-pb-lg">
               <img class src="statics/Logo/logo" width="100" />
             </div>
-            <div>
-              <form
-                v-if="!CONFIG.IS_BLOCKED_IP && showLoginForm"
-                @submit.prevent.stop="onSubmit"
-                class="col q-gutter-md q-mx-lg"
-                :autocomplete="CONFIG.LOGIN_PAGE_REMEMBER_CREDENTIALS ? 'on' : 'off'"
-              >
-                <q-input
-                  type="text"
-                  v-model="user"
-                  :label="$t('LBL_USER')"
-                  lazy-rules
-                  :rules="[val => (val && val.length > 0) || 'Please type something']"
-                  :autocomplete="CONFIG.LOGIN_PAGE_REMEMBER_CREDENTIALS ? 'on' : 'off'"
-                />
-                <q-input
-                  type="password"
-                  v-model="password"
-                  :label="$t('Password')"
-                  lazy-rules
-                  :rules="[val => (val && val.length > 0) || 'Please type something']"
-                  :autocomplete="CONFIG.LOGIN_PAGE_REMEMBER_CREDENTIALS ? 'on' : 'off'"
-                />
-                <q-select
-                  v-if="CONFIG.LANGUAGE_SELECTION"
-                  v-model="language"
-                  :options="CONFIG.LANGUAGES"
-                  :label="$t('LBL_CHOOSE_LANGUAGE')"
-                >
-                  <template v-slot:prepend>
-                    <q-icon name="translate" />
-                  </template>
-                </q-select>
-                <q-select
-                  v-if="CONFIG.LAYOUT_SELECTION"
-                  v-model="layout"
-                  :options="CONFIG.LAYOUTS"
-                  :label="$t('LBL_SELECT_LAYOUT')"
-                >
-                  <template v-slot:prepend>
-                    <q-icon name="looks" />
-                  </template>
-                </q-select>
-                <q-btn
-                  size="lg"
-                  :label="$t('LBL_SIGN_IN')"
-                  type="submit"
-                  color="secondary"
-                  class="full-width q-mt-lg"
-                />
-                <a
-                  v-if="CONFIG.FORGOT_PASSWORD"
-                  @click="
-                    showRemindForm = !showRemindForm
-                    showLoginForm = !showLoginForm
-                  "
-                  class="text-secondary float-right"
-                  href="#"
-                  >{{ $t('ForgotPassword') }}?</a
-                >
-              </form>
-              <q-banner v-else-if="CONFIG.IS_BLOCKED_IP" class="bg-negative q-mt-lg text-white">
-                <div class="text-center">
-                  <q-icon name="remove_circle_outline" class="text-h1 q-pb-md"></q-icon>
-                </div>
-                <p>{{ $t('LBL_IP_IS_BLOCKED') }}</p>
-              </q-banner>
-              <q-banner v-if="CONFIG.MESSAGE" :class="[msgClass, 'q-mt-lg', 'text-white']">
-                <p>{{ CONFIG.MESSAGE }}</p>
-              </q-banner>
-            </div>
-            <template v-if="CONFIG.FORGOT_PASSWORD && showRemindForm">
-              <form class="col q-gutter-md q-mx-lg" action="index.php?module=Users&action=ForgotPassword" method="POST">
-                <q-input
-                  type="text"
-                  v-model="reminderUser"
-                  :title="$t('LBL_USER')"
-                  :label="$t('LBL_USER')"
-                  autocomplete="off"
-                >
-                  <template v-slot:prepend>
-                    <q-icon name="person" />
-                  </template>
-                </q-input>
-                <q-input
-                  type="text"
-                  v-model="reminderEmail"
-                  autocomplete="off"
-                  :title="$t('LBL_EMAIL')"
-                  :label="$t('LBL_EMAIL')"
-                >
-                  <template v-slot:prepend>
-                    <q-icon name="mail_outline" />
-                  </template>
-                </q-input>
-                <q-btn size="lg" :label="$t('LBL_SEND')" type="submit" color="secondary" class="full-width q-mt-lg" />
-                <a
-                  @click="
-                    showLoginForm = !showLoginForm
-                    showRemindForm = !showRemindForm
-                  "
-                  href="#"
-                  class="text-secondary float-right"
-                  >{{ $t('LBL_TO_CRM') }}</a
-                >
-              </form>
-            </template>
+            <keep-alive>
+              <component
+                v-if="!CONFIG.IS_BLOCKED_IP"
+                :is="activeComponent"
+                :CONFIG="CONFIG"
+                :toggleActiveComponent="toggleActiveComponent"
+              />
+            </keep-alive>
+            <q-banner v-if="CONFIG.IS_BLOCKED_IP" class="bg-negative q-mt-lg text-white">
+              <p>{{ $t('LBL_IP_IS_BLOCKED') }}</p>
+            </q-banner>
+            <q-banner v-if="CONFIG.MESSAGE" :class="[msgClass, 'q-mt-lg', 'text-white']">
+              <p>{{ CONFIG.MESSAGE }}</p>
+            </q-banner>
           </div>
         </div>
       </q-page>
@@ -129,7 +36,8 @@
 </style>
 
 <script>
-import actions from '../store/actions.js'
+import Form from '../components/Login/Form.vue'
+import Reminder from '../components/Login/Reminder.vue'
 const CONFIG = {
   // component config loaded from server
   LANGUAGES: ['polish', 'english', 'german'],
@@ -147,35 +55,19 @@ export default {
   name: 'Login',
   data() {
     return {
-      user: '',
-      password: '',
-      language: CONFIG.DEFAULT_LANGUAGE, //AppConfig::main('default_language')
-      layout: '',
-      showRemindForm: false,
+      activeComponent: 'login-form',
+      showReminderForm: false,
       showLoginForm: true,
-      reminderEmail: '',
-      reminderUser: '',
       CONFIG: CONFIG
     }
   },
-  computed: {
-    msgClass: function() {
-      return {
-        'bg-positive': this.CONFIG.MESSAGE_TYPE === 'success',
-        'bg-negative': this.CONFIG.MESSAGE_TYPE === 'error',
-        'bg-warning': this.CONFIG.MESSAGE_TYPE === ''
-      }
-    }
+  components: {
+    loginForm: Form,
+    reminderForm: Reminder
   },
   methods: {
-    onSubmit() {
-      this.$refs.user.validate()
-      this.$refs.password.validate()
-      if (this.$refs.user.hasError || this.$refs.password.hasError) {
-        this.formHasError = true
-      } else {
-        this.$store.dispatch(actions.Login.login, { user: this.user, password: this.password })
-      }
+    toggleActiveComponent: function(componentName) {
+      this.activeComponent = componentName
     }
   }
 }
